confirm before deleting folders

+++ components/listOfNotes.js
@@ -134,27 +134,31 @@ export default function ListOfNotes({ inApp }) {
     
     // Same function as removeNote() but without the router.push because you can't edit the folder.
     async function removeFolder(id) {
-        for (let i=0; i<folderData.length; i++) {
-            if (folderData[i].id == id) {
-                for (let j=0; j<folderData[i].notes.length; j++) {
-                    const { dataNote, errorNote } = await supabase
-                        .from('notesv2')
-                        .update({inFolder: false})
-                        .eq('id', folderData[i].notes[j].id)
+        let text = "Are you sure?\n\nDeleting this folder will delete all its notes too.\nRemove the notes you want to keep from the folder first!"
+        if (confirm(text)) {
+            for (let i=0; i<folderData.length; i++) {
+                if (folderData[i].id == id) {
+                    for (let j=0; j<folderData[i].notes.length; j++) {
+                        const { dataNote, errorNote } = await supabase
+                            .from('notesv2')
+                            .delete()
+                            .eq('id', folderData[i].notes[j].id)
+                    }
                 }
             }
+            const { data, error } = await supabase
+                .from('folders')
+                .delete()
+                .eq('id', id)
+            if (data) {
+                //array of todos
+                setFolderData(data)
+            } if (error) {
+                console.log('error', error)
+                return
+            }
         }
-        const { data, error } = await supabase
-            .from('folders')
-            .delete()
-            .eq('id', id)
-        if (data) {
-            //array of todos
-            setFolderData(data)
-        } if (error) {
-            console.log('error', error)
-            return
-        }
+
         getNotes()
         getFolders()
     }
